+++ packages/integration-sdk-entity-validator/src/index.ts
@@ -1,0 +1,6 @@
+export * from './validator';
+export {
+  unknownPropertySymbol,
+  isEntityValidationError,
+  type EntityValidationError,
+} from './entityValidationError';